# contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
...
info:
  title: Talosmine Control Plane API
  version: 1.0.0
  description: |
    Control Plane của Talosmine — nguồn sự thật về identity, entitlement và quota.

    **Trạng thái: P1 (bootstrap).** Chỉ có health endpoint. Không có capability nghiệp vụ nào.

    Quy ước xuyên phase:
    - Số lượng (quantity/limit/counter/amount) biểu diễn bằng **decimal string**, không phải
      JSON number — PostgreSQL dùng `bigint` và JSON number mất chính xác ngoài 2^53.
    - Thời gian dùng timestamp có timezone; business time dựa DB clock.
    - Mọi lỗi trả về `ErrorEnvelope` với `code` máy đọc được. Client bắt theo `code`,
      không theo `message`.
  license:
    name: UNLICENSED
servers:
  - url: https://talosmine.localhost
    description: Local dev qua Caddy (DEC-T11)
...
paths:
...
  /v1/auth/sessions:
    post:
      tags: [auth]
      operationId: exchangeIdTokenForSession
      summary: Đổi ID token của IdP lấy phiên Talosmine
      description: |
        Cầu nối giữa OIDC và phiên nội bộ. BFF chạy luồng Authorization Code với IdP rồi
        gửi NGUYÊN `id_token` tới đây.

        Control Plane **tự xác minh chữ ký** bằng JWKS của issuer — nó không nhận
        `issuer`/`subject`/`accountId` do caller khai. Nếu tin claim của caller thì bất kỳ
        ai gọi được endpoint này đều mạo danh được người khác.

        Sau khi xác minh: tìm hoặc tạo account theo `(issuer, subject)`, kiểm trạng thái
        account, rồi phát session token. Endpoint này KHÔNG cần phiên có sẵn.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              additionalProperties: false
              properties:
                idToken:
                  type: string
                  description: ID token JWT do IdP phát.
      responses:
        '201':
          description: Đã tạo phiên.
          content:
            application/json:
              schema:
                type: object
                required: [sessionToken, csrfToken, expiresAt, accountId, created]
                properties:
                  sessionToken:
                    type: string
                    description: Token thô. Chỉ trả về MỘT LẦN; DB chỉ lưu hash.
                  csrfToken:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                  accountId:
                    type: string
                    format: uuid
                  created:
                    type: boolean
                    description: true nếu account vừa được tạo lần đầu.
        '400':
          description: Thiếu `idToken`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token không hợp lệ — sai chữ ký, hết hạn, sai issuer hoặc sai audience.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: |
            Account không ở trạng thái `active`. Thông điệp cố ý chung chung, không phân
            biệt `pending` với `disabled`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /v1/me/account/sessions/all:
    delete:
      tags: [account]
      operationId: revokeAllOwnSessions
      summary: Đăng xuất mọi nơi
      description: |
        Thu hồi TẤT CẢ phiên còn hiệu lực của account, kể cả phiên đang gọi request này.
        Idempotent: gọi lại khi không còn phiên nào vẫn trả 204.
      security:
        - webSession: []
      responses:
        '204':
          description: Đã thu hồi (không có nội dung trả về).
        '401':
          description: Không có phiên hợp lệ.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  securitySchemes:
    webSession:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Web session token do BFF forward (từ cookie HttpOnly). Control Plane tra hash
        trong `web_sessions`, kiểm chưa hết hạn/chưa thu hồi, rồi resolve accountId.
        Token thô KHÔNG được log. Đây là kênh của NGƯỜI DÙNG — khác `serviceIdentity` (M2M).
    userSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token user do Auth0 phát, được BFF chuyển tiếp.
        **CHƯA dùng ở P1** — wiring Auth0 thật là P2 và phụ thuộc DEC-B03 (tenant thật).
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        M2M token của backend ứng dụng (Data Plane). Mỗi backend có identity RIÊNG,
        audience cụ thể và exact resource scope. Không dùng chung credential.
        **CHƯA dùng ở P1.**
  schemas:
...
    ErrorEnvelope:
      type: object
      description: |
        Hình dạng lỗi ổn định cho MỌI lỗi rời khỏi API, kể cả lỗi không lường trước.

        Ràng buộc bảo mật: không bao giờ chứa stack trace, secret, connection string
        hay chi tiết nội bộ. Với lỗi 5xx, `message` luôn là thông điệp chung — chi tiết
        thật chỉ đi vào log phía server, truy được bằng `correlationId`.
      required: [code, message, correlationId]
      additionalProperties: false
      properties:
        code:
          type: string
          description: |
            Mã máy đọc được, ổn định, không phụ thuộc câu chữ giao diện.
            Client bắt lỗi theo trường này.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - INTERNAL_ERROR
            - VALIDATION_FAILED
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Thông điệp an toàn để hiển thị.
          maxLength: 500
        correlationId:
          type: string
          format: uuid
          description: |
            Truy vết xuyên BFF -> API -> worker -> log.
            Nhận từ header `X-Correlation-Id` nếu hợp lệ, ngược lại server sinh mới (UUIDv7).
            **Không phải** idempotency key — đừng dùng lẫn.
        details:
          type: object
          description: Chi tiết bổ sung theo allowlist. Không bao giờ chứa dữ liệu nhạy cảm.
          additionalProperties: true
...
security: []